cleaned up parser and lexer code to reduce high coupling
