+++ MathsVisualisationTool/interpreter/Parser.cs
@@ -16,9 +16,6 @@ namespace MathsVisualisationTool
         private Token nextToken = null;
         //the index of the NEXT token to be analysed when getNextToken() is called.
         private int index = 0;
-        //bracketLevel is for storing the bracket level of an expression
-        // i.e. ( will set the bracket Level to 1 but a ) will set it back to 0 again.
-        private int bracketLevel = 0;
         
         public Parser()
         {
@@ -33,13 +30,27 @@ namespace MathsVisualisationTool
         public double AnalyseTokens(List<Token> tokens)
         {
             this.tokens = tokens;
+
+            //Analyse the syntax to see if it is valid.
+            SyntaxAnalyser s = new SyntaxAnalyser(tokens);
+            s.checkTokens();
+
+            //If it is then do some preprocessing on the input.
+            Preprocessor p = new Preprocessor(tokens);
+            this.tokens = p.processTokens();
+
             getNextToken();
 
             double value = analyseExpressions(double.NaN);
             return value;
         }
 
-
+        /// <summary>
+        /// Method called to process the expression. It assumes that it is in a universal format
+        /// as the checking is done by the preprocessor and syntax analyser.
+        /// </summary>
+        /// <param name="value"></param>
+        /// <returns></returns>
         public double analyseExpressions(double value)
         {
             while (nextToken != null)
@@ -47,31 +58,29 @@ namespace MathsVisualisationTool
                 SUPPORTED_TOKENS tokenType = nextToken.GetType();
                 if (tokenType == SUPPORTED_TOKENS.INTEGER)
                 {
+                    //If the next token is an int, then just collect the value
                     value = Convert.ToDouble(nextToken.GetValue());
                     getNextToken();
                 }
                 else if (tokenType == SUPPORTED_TOKENS.DIVISION || tokenType == SUPPORTED_TOKENS.MULTIPLICATION)
                 {
+                    //If its division or multiplication, run the appropriate method.
                     value = divisionAndMultHandle(value);
-                    getNextToken();
                 }
                 else if (tokenType == SUPPORTED_TOKENS.PLUS || tokenType == SUPPORTED_TOKENS.MINUS)
                 {
+                    //same for plus and minus.
                     value = plusAndMinusHandle(value);
                 }
                 else if (tokenType == SUPPORTED_TOKENS.OPEN_BRACKET)
                 {
-                    bracketLevel++;
+                    //if its an open bracket then recursively call on the expression encapsulated inside
+                    //the brackets.
                     getNextToken();
                     value = analyseExpressions(value);
                 } else if (tokenType == SUPPORTED_TOKENS.CLOSE_BRACKET)
                 {
-                    //Cannot have a closing bracket without an open bracket.
-                    if(bracketLevel == 0)
-                    {
-                        throw new Exception("Unrecognised character ')' at token position " + index);
-                    }
-                    bracketLevel--;
+                    //if its a close bracket then just return.
                     getNextToken();
                     return value;
                 }
@@ -91,70 +100,38 @@ namespace MathsVisualisationTool
             Expression right;
             SUPPORTED_TOKENS op = nextToken.GetType();
 
-            //so there is nothing before 
-            if (double.IsNaN(value))
-            {
-                throw new Exception("Integer expected at token position " + (index - 1));
-            }
-
             //Find the left and right hand side of the expression.
             left = new Constant(value);
 
-            Token peekToken = peek();
+            //get the next token
+            getNextToken();
 
-            //check what the next token is as it should be either a plus, minus or and int.
-            if(peekToken == null || 
-                !(peekToken.GetType() == SUPPORTED_TOKENS.PLUS) 
-                && !(peekToken.GetType() == SUPPORTED_TOKENS.MINUS)
-                && !(peekToken.GetType() == SUPPORTED_TOKENS.INTEGER)
-                && !(peekToken.GetType() == SUPPORTED_TOKENS.OPEN_BRACKET))
+            double rightValue = 0.0;
+
+            //If a bracket has been found then analyse the expression inside the bracket first.
+            if (nextToken.GetType() == SUPPORTED_TOKENS.OPEN_BRACKET)
             {
-                throw new Exception("Integer expected at token position " + index);
+                getNextToken();
+                rightValue = analyseExpressions(value);
             } else
             {
-                //get the next token
-                getNextToken();
-            
-                double multiplier = 1;
-
-                //gather the multiplier for situations like 8*-8 = -64 rather than 64.
-                while(nextToken.GetType() == SUPPORTED_TOKENS.PLUS || nextToken.GetType() == SUPPORTED_TOKENS.MINUS)
-                {
-                    if(nextToken.GetType() == SUPPORTED_TOKENS.PLUS)
-                    {
-                        multiplier *= 1;
-                    } else
-                    {
-                        multiplier *= -1;
-                    }
-                    getNextToken();
-                }
-
-                double rightValue = 0.0;
-
-                //If a bracket has been found then analyse the expression inside the bracket first.
-                if (nextToken.GetType() == SUPPORTED_TOKENS.OPEN_BRACKET)
-                {
-                    bracketLevel++;
-                    getNextToken();
-                    rightValue = multiplier*analyseExpressions(value);
-                } else
-                {
-                    rightValue = multiplier * Convert.ToDouble(nextToken.GetValue());
-                }
-
-                right = new Constant(rightValue);
-
-                //Prevent division by zero
-                if (rightValue == 0 && op == SUPPORTED_TOKENS.DIVISION)
-                {
-                    throw new Exception("Cannot divide by zero");
-                }
-
-                Expression ne = new Operation(left, op, right);
-
-                return ne.Evaluate();
+                //else its just a number so extract the value of it.
+                rightValue = Convert.ToDouble(nextToken.GetValue());
             }
+
+            right = new Constant(rightValue);
+
+            //Prevent division by zero
+            if (rightValue == 0 && op == SUPPORTED_TOKENS.DIVISION)
+            {
+                throw new Exception("Cannot divide by zero");
+            }
+
+            Expression ne = new Operation(left, op, right);
+
+            getNextToken();
+
+            return ne.Evaluate();
         }
 
         /// <summary>
@@ -168,44 +145,23 @@ namespace MathsVisualisationTool
             Expression right = null;
             SUPPORTED_TOKENS op = nextToken.GetType();
 
-            //so there is nothing before as a situation like '+3' is allowed in the language. 
-            if (double.IsNaN(value))
-            {
-                left = new Constant(0);
-            } else
-            {
-                left = new Constant(value);
-            }
+            //Get the left hand side of the expression.
+            left = new Constant(value);
 
-            Token peekToken = peek();
-            double multiplier = 1;
+            //Now look at the token after the operator
+            getNextToken();
 
-            while(!(peekToken == null))
+            if(nextToken.GetType() == SUPPORTED_TOKENS.INTEGER)
             {
-                if(peekToken.GetType() == SUPPORTED_TOKENS.INTEGER)
-                {
-                    //found a number so stop.
-                    right = new Constant(Convert.ToDouble(peekToken.GetValue()) * multiplier);
-                    //2 next token calls to skip to the next token after the integer token.
-                    getNextToken();
-                    getNextToken();
-                    break;
-                } else if (peekToken.GetType() == SUPPORTED_TOKENS.OPEN_BRACKET)
-                {
-                    getNextToken();
-                    bracketLevel++;
-                    right = new Constant(analyseExpressions(value));
-                    break;
-                } else 
-                {
-                    throw new Exception("Invalid token at position - " + index);
-                }
-            }
-
-            //so if there is no right hand expression.
-            if(right == null)
+                //A number so just extract the value from it.
+                right = new Constant(Convert.ToDouble(nextToken.GetValue()));
+                //Then call get next token.
+                getNextToken();
+            } else if (nextToken.GetType() == SUPPORTED_TOKENS.OPEN_BRACKET)
             {
-                throw new Exception("Integer expected at position - " + (index));
+                getNextToken();
+                //recursively call to analyse the expression enclosed in brackets.
+                right = new Constant(analyseExpressions(value));
             }
 
             Expression ne = new Operation(left, op, right);
