+++ TODO.txt
@@ -1,1 +1,1 @@
-Test with custom parenthesis. - Sort of works situations like (2+(2+(3*4))) it chops off the last two parentheses.
+Works well when input is valid - write the syntax analyser class for this.
